refactor: group host acp modules
