+++ packages/browseros-agent/apps/server/src/lib/agents/adapter-health.ts
@@ -4,13 +4,13 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
+import type { AgentAdapter } from './agent-types'
+import { isHostAcpAdapter } from './host-acp/config'
 import {
   type AdapterHealth,
   type DetectHostAdapterOptions,
   detectHostAdapter,
-} from './adapter-detection'
-import type { AgentAdapter } from './agent-types'
-import { isHostAcpAdapter } from './host-acp-adapter-config'
+} from './host-acp/detection'
 import {
   type AgentRuntime,
   type AgentRuntimeRegistry,
