+++ packages/browseros-agent/apps/server/src/lib/agents/host-acp/binary-resolver.ts
